--- com.unity.netcode.gameobjects/Runtime/Messaging/Messages/MessageTypes.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1fbafb5a0a1ab924eb0bdb1870f45d97
