+++ com.unity.netcode.gameobjects/Tests/Editor/Messaging/MessageRegistrationTests.cs
@@ -191,77 +191,5 @@ namespace Unity.Netcode.EditorTests
                 Assert.AreEqual(handlerFour, systemThree.MessageHandlers[systemThree.GetMessageType(typeof(TestMessageFour))]);
             }
         }
-
-        internal class AAAEarlyLexicographicNetworkMessage : INetworkMessage
-        {
-            public void Serialize(FastBufferWriter writer, int targetVersion)
-            {
-            }
-
-            public bool Deserialize(FastBufferReader reader, ref NetworkContext context, int receivedMessageVersion)
-            {
-                return true;
-            }
-
-            public void Handle(ref NetworkContext context)
-            {
-            }
-
-            public int Version => 0;
-        }
-
-#pragma warning disable IDE1006
-        internal class zzzLateLexicographicNetworkMessage : AAAEarlyLexicographicNetworkMessage
-        {
-        }
-#pragma warning restore IDE1006
-
-        internal class OrderingMessageProvider : INetworkMessageProvider
-        {
-            public List<NetworkMessageManager.MessageWithHandler> GetMessages()
-            {
-                var listMessages = new List<NetworkMessageManager.MessageWithHandler>();
-
-                var messageWithHandler = new NetworkMessageManager.MessageWithHandler
-                {
-                    MessageType = typeof(zzzLateLexicographicNetworkMessage),
-                    GetVersion = NetworkMessageManager.CreateMessageAndGetVersion<zzzLateLexicographicNetworkMessage>
-                };
-                listMessages.Add(messageWithHandler);
-
-                messageWithHandler.MessageType = typeof(ConnectionRequestMessage);
-                messageWithHandler.GetVersion = NetworkMessageManager.CreateMessageAndGetVersion<ConnectionRequestMessage>;
-                listMessages.Add(messageWithHandler);
-
-                messageWithHandler.MessageType = typeof(ConnectionApprovedMessage);
-                messageWithHandler.GetVersion = NetworkMessageManager.CreateMessageAndGetVersion<ConnectionApprovedMessage>;
-                listMessages.Add(messageWithHandler);
-
-                messageWithHandler.MessageType = typeof(AAAEarlyLexicographicNetworkMessage);
-                messageWithHandler.GetVersion = NetworkMessageManager.CreateMessageAndGetVersion<AAAEarlyLexicographicNetworkMessage>;
-                listMessages.Add(messageWithHandler);
-
-                return listMessages;
-            }
-        }
-
-        [Test]
-        public void MessagesGetPrioritizedCorrectly()
-        {
-            var sender = new NopMessageSender();
-            var provider = new OrderingMessageProvider();
-            using var messageManager = new NetworkMessageManager(sender, null, provider);
-
-            // the 2 priority messages should appear first, in lexicographic order
-            Assert.AreEqual(messageManager.MessageTypes[0], typeof(ConnectionApprovedMessage));
-            Assert.AreEqual(messageManager.MessageTypes[1], typeof(ConnectionRequestMessage));
-
-            // the other should follow after
-            Assert.AreEqual(messageManager.MessageTypes[2], typeof(AAAEarlyLexicographicNetworkMessage));
-            Assert.AreEqual(messageManager.MessageTypes[3], typeof(zzzLateLexicographicNetworkMessage));
-
-            // there should not be any extras
-            Assert.AreEqual(messageManager.MessageHandlerCount, 4);
-        }
     }
 }
