+++ com.unity.netcode.gameobjects/Tests/Editor/Transports/UnityTransportTests.cs
@@ -10,13 +10,13 @@ namespace Unity.Netcode.EditorTests
         [SetUp]
         public void OnSetup()
         {
-            NetworkMessageManager.IntegrationTest = true;
+            ILPPMessageProvider.IntegrationTestNoMessages = true;
         }
 
         [TearDown]
         public void OnTearDown()
         {
-            NetworkMessageManager.IntegrationTest = false;
+            ILPPMessageProvider.IntegrationTestNoMessages = false;
         }
 
         // Check that starting an IPv4 server succeeds.
